Move GenerateNanorc to gen folder
